Split PiP callback implementation for lint
